move timer to own module
