+++ src/newtab/index.js
@@ -7,7 +7,7 @@ import './store';
 import { initEventsHandler } from './events-handler';
 import './background-image';
 import './todolist/todolist';
-import { initTimer } from './timer';
+import { initTimer } from './timer/timer';
 
 initEventsHandler();
 initTimer();
